+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+#### Added
+
+- [#1301](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1301) Add support for INDEX INCLUDE.
+
 #### Changed
 
 - [#1273](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1273) TinyTDS v3+ is now required.